+++ app/src/main/res/drawable/ic_walletconnect.xml
@@ -1,0 +1,13 @@
+<vector
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="18dp"
+    android:height="11dp"
+    android:viewportHeight="185"
+    android:viewportWidth="300">
+    <path
+        android:fillColor="@color/battleship_grey"
+        android:fillType="nonZero"
+        android:pathData="M61.439,36.256C110.35,-11.632 189.651,-11.632 238.562,36.256L244.448,42.02C246.894,44.414 246.894,48.296 244.448,50.691L224.312,70.406C223.089,71.603 221.106,71.603 219.884,70.406L211.783,62.475C177.661,29.067 122.339,29.067 88.217,62.475L79.542,70.969C78.32,72.166 76.337,72.166 75.114,70.969L54.978,51.253C52.532,48.859 52.532,44.977 54.978,42.582L61.439,36.256ZM280.206,77.03L298.128,94.577C300.574,96.971 300.574,100.853 298.128,103.248L217.318,182.369C214.872,184.763 210.907,184.763 208.462,182.369C208.462,182.369 208.462,182.369 208.462,182.369L151.108,126.214C150.496,125.616 149.505,125.616 148.894,126.214C148.894,126.214 148.894,126.214 148.894,126.214L91.541,182.369C89.095,184.763 85.13,184.763 82.684,182.369C82.684,182.369 82.684,182.369 82.684,182.369L1.872,103.247C-0.574,100.852 -0.574,96.97 1.872,94.576L19.794,77.029C22.239,74.635 26.204,74.635 28.65,77.029L86.005,133.184C86.616,133.783 87.607,133.783 88.219,133.184C88.219,133.184 88.219,133.184 88.219,133.184L145.571,77.029C148.017,74.635 151.982,74.634 154.427,77.029C154.427,77.029 154.427,77.029 154.427,77.029L211.782,133.184C212.394,133.783 213.385,133.783 213.996,133.184L271.35,77.03C273.796,74.636 277.761,74.636 280.206,77.03Z"
+        android:strokeColor="#00000000"
+        android:strokeWidth="1" />
+</vector>
