+++ app/src/main/res/layout/layout_safe_main.xml
@@ -286,13 +286,29 @@
                     app:layout_constraintTop_toBottomOf="@id/layout_safe_main_tokens"
                     tools:drawableStart="@drawable/ic_settings_address_book" />
 
+                <!-- Container to hide this when not in Debug Mode (else it is controlled by ContraintGroup) -->
+                <FrameLayout
+                    android:id="@+id/layout_safe_main_wallet_connect_container"
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"
+                    app:layout_constraintTop_toBottomOf="@id/layout_safe_main_address_book">
+
+                    <TextView
+                        android:id="@+id/layout_safe_main_wallet_connect"
+                        style="@style/NavigationText"
+                        android:layout_width="match_parent"
+                        android:layout_height="wrap_content"
+                        android:text="@string/wallet_connect"
+                        tools:drawableStart="@drawable/ic_walletconnect" />
+                </FrameLayout>
+
                 <TextView
                     android:id="@+id/layout_safe_main_general_settings"
                     style="@style/NavigationText"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:text="@string/settings"
-                    app:layout_constraintTop_toBottomOf="@id/layout_safe_main_address_book"
+                    app:layout_constraintTop_toBottomOf="@id/layout_safe_main_wallet_connect_container"
                     tools:drawableStart="@drawable/ic_general_settings" />
 
 
@@ -340,7 +356,7 @@
                     android:id="@+id/layout_safe_main_navigation_settings"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-                    app:constraint_referenced_ids="layout_safe_main_tokens,layout_safe_main_general_settings,layout_safe_main_address_book,layout_safe_main_account_container,layout_safe_main_debug_settings_container"
+                    app:constraint_referenced_ids="layout_safe_main_tokens,layout_safe_main_general_settings,layout_safe_main_address_book,layout_safe_main_wallet_connect_container,layout_safe_main_account_container,layout_safe_main_debug_settings_container"
                     tools:visibility="visible" />
             </androidx.constraintlayout.widget.ConstraintLayout>
         </androidx.core.widget.NestedScrollView>
