+++ app/src/main/res/layout/layout_wallet_connect_sessions_item.xml
@@ -67,7 +67,7 @@
         android:background="@drawable/selectable_background"
         android:gravity="center"
         android:padding="16dp"
-        android:text="Deny"
+        android:text="@string/deny"
         android:visibility="gone"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toStartOf="@+id/layout_wallet_connect_sessions_item_kill"
@@ -81,7 +81,7 @@
         android:background="@drawable/selectable_background"
         android:gravity="center"
         android:padding="16dp"
-        android:text="Approve"
+        android:text="@string/approve"
         android:visibility="gone"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toStartOf="@+id/layout_wallet_connect_sessions_item_deny"
@@ -95,7 +95,7 @@
         android:background="@drawable/selectable_background"
         android:gravity="center"
         android:padding="16dp"
-        android:text="Activate"
+        android:text="@string/activate"
         android:visibility="gone"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toStartOf="@+id/layout_wallet_connect_sessions_item_approve"
