+++ app/src/main/res/values/strings.xml
@@ -265,4 +265,9 @@ Without it, you won’t be able to recover your Safe and money.</string>
     <string name="review_message">Review message</string>
     <string name="sign_message_notification_title">Sign Message</string>
     <string name="sign_message_notification_description">You have been requested to sign a message</string>
+    <string name="wallet_connect">WalletConnect</string>
+    <string name="wallet_connect_sessions">WalletConnect Sessions</string>
+    <string name="deny">Deny</string>
+    <string name="approve">Approve</string>
+    <string name="activate">Activate</string>
 </resources>
